% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grcm39_chrlen.R
\name{grcm39_chrlen}
\alias{grcm39_chrlen}
\title{Mouse chromosome lengths in basepairs for build GRCm39}
\source{
\url{https://www.ncbi.nlm.nih.gov/datasets/genome/GCF_000001635.27/}
}
\description{
A vector of mouse chromosome lengths in basepairs for mouse genome build GRCm39.
}
\examples{
data(grcm39_chrlen)
}
\keyword{datasets}
