% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmconvert-package.R
\docType{package}
\name{mmconvert-package}
\alias{mmconvert-package}
\title{mmconvert: Mouse Map Converter}
\description{
Function to convert mouse genome positions between the build 39
physical map and the Cox genetic map
\doi{10.1534/genetics.109.105486}.
}
\section{Useful links}{

\itemize{
\item Source at \url{https://github.com/rqtl/mmconvert}
\item BugReports to \url{https://github.com/rqtl/mmconvert/issues}
}
}

\keyword{internal}
